Move connection req error handling to self connection
It's now done by handler for normal connections.
