+++ node/connection_base.js
@@ -91,33 +91,14 @@ function connBaseRequest(options) {
 
 TChannelConnectionBase.prototype.handleCallRequest = function handleCallRequest(req) {
     var self = this;
-    
-    self.ops.addInReq(req);
 
     req.remoteAddr = self.remoteName;
-    req.errorEvent.on(onReqError);
+    self.ops.addInReq(req);
 
     process.nextTick(runHandler);
 
-    function onReqError(err) {
-        self.onReqError(req, err);
-    }
-
     function runHandler() {
         self.runHandler(req);
-    }
-};
-
-TChannelConnectionBase.prototype.onReqError = function onReqError(req, err) {
-    var self = this;
-    if (!req.res) self.buildResponse(req, {});
-    if (err.type === 'tchannel.timeout' ||
-        err.type === 'tchannel.request.timeout'
-    ) {
-        req.res.sendError('Timeout', err.message);
-    } else {
-        var errName = err.name || err.constructor.name;
-        req.res.sendError('UnexpectedError', errName + ': ' + err.message);
     }
 };
